+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## 0.4.0
+- Markdown as hovering string.
+- More unit tests.
+
+## 0.3.0
+- Adding flags and decriptions.
+
 ## 0.2.0
 - Added Z80N instructions.
-
 
 ## 0.1.0
 - Initial version. Mainly copied from asm-code-lens.